refactor(sidebar): extract row/tip/noise views into SidebarComponents.swift
Lifts four fileprivate NSView subclasses out of SidebarView.swift:
SidebarTabRow, SidebarToolRow, SidebarNoiseView, WorkspaceTipView.

None of them reference SidebarView state — they communicate via closures
already (onSelect / onClose / etc.). fileprivate drops to internal so
SidebarView's cross-file references resolve.

SidebarView.swift: 1488 → 855 lines (-42.5%).
